+++ src/unix/processx.c
@@ -485,7 +485,6 @@ SEXP processx_exec(SEXP command, SEXP args, SEXP std_in, SEXP std_out,
     processx_connection_t *ccon =
       R_ExternalPtrAddr(VECTOR_ELT(connections, i));
     int fd = processx_c_connection_fileno(ccon);
-    processx__nonblock_fcntl(fd, 0);
     pipes[i + 3][1] = fd;
   }
 
@@ -575,13 +574,6 @@ SEXP processx_exec(SEXP command, SEXP args, SEXP std_in, SEXP std_out,
   /* Closed unused ends of pipes */
   for (i = 0; i < 3; i++) {
     if (pipes[i][1] >= 0) close(pipes[i][1]);
-  }
-
-  /* Close connections passed to the child process */
-  for (i = 3; i < num_connections; i++) {
-    processx_connection_t *ccon =
-      R_ExternalPtrAddr(VECTOR_ELT(connections, i - 3));
-    processx_c_connection_close(ccon);
   }
 
   /* Create proper connections */